Cleaning warning, leaving enclave as debug for licensing gods sake

+++ secureUI/encryptedfile.cpp
@@ -32,9 +32,9 @@ bool EncryptedFile::setFileName(char*fname)
 {
 	strcpy(m_testfname, fname);
 	m_testFile = fopen(m_testfname, "rb");
-	size_t fpos = ftell(m_testFile);
+	long fpos = ftell(m_testFile);
 	fseek(m_testFile, 0, SEEK_END);
-	size_t fsize = ftell(m_testFile);
+	long fsize = ftell(m_testFile);
 	fseek(m_testFile, fpos, SEEK_SET);
 	m_fsize = fsize;
 	return true;
